requirements.txt
streamlit==1.26.0
google-generativeai>=0.7.0,<0.8.0
langchain-google-genai==1.0.10
python-dotenv==1.0.0
